time to break a lot of things
