+++ freesprite/EditorColorPicker.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "drawable.h"
 #include "DrawableManager.h"
+#include "UIColorPicker.h"
 #include "EventCallbackListener.h"
 #include "UIHueSlider.h"
 #include "UIColorSlider.h"
@@ -12,60 +13,16 @@
 #include "DraggablePanel.h"
 #include "colormodels.h"
 
-struct ColorModelValue {
-    UIColorSlider* valueSlider;
-    UILabel* valueLabel;
-    double valueNow;
-    std::pair<double, double> range;
-};
-struct ColorModelData {
-    ColorModel* targetModel;
-    std::map<std::string, ColorModelValue> components;
-};
-class ColorPickerColorButton : public UIButton {
-public:
-    EditorColorPicker* parent = NULL;
-    u32 color = 0;
 
-    ColorPickerColorButton(EditorColorPicker* parent, u32 color);
-    void click() override;
-};
-
-class EditorColorPicker : public DraggablePanel, public EventCallbackListener
+class EditorColorPicker : public UIColorPicker
 {
 protected:
-    EditorColorPicker() {}
+	EditorColorPicker() : UIColorPicker(400, 390) {}
 public:
-    MainEditor* caller;
-
-    double currentH = 0, currentS = 0, currentV = 0;
-
-    uint8_t currentR = 0, currentG = 0, currentB = 0;
-
-    UIHueSlider* hueSlider = NULL;
-    UISVPicker* satValSlider = NULL;
-    UITextField* colorTextField = NULL;
-
-
-    //todo: change these sliders to work based off the color models system
-    UITextField* txtR, *txtG, *txtB;
-    UITextField* txtH, *txtS, *txtV;
-
-    UIColorSlider* sliderH = NULL;
-    UIColorSlider* sliderS = NULL;
-    UIColorSlider* sliderV = NULL;
-    UIColorSlider* sliderR = NULL;
-    UIColorSlider* sliderG = NULL;
-    UIColorSlider* sliderB = NULL;
-
-    TabbedView* colorModeTabs = NULL;
-    TabbedView* colorTabs = NULL;
-    ScrollingPanel* palettePanel = NULL;
+    MainEditor* caller = NULL;
 
     UIButton* eraserButton = NULL;
     UIButton* blendModeButton = NULL;
-
-    std::vector<std::pair<std::string, ColorModelData>> colorModels;
 
     std::vector<uint32_t> lastColors;
     bool lastColorsChanged = true;
@@ -74,43 +31,11 @@ public:
 
     void render(XY position) override;
 
-    void eventTextInput(int evt_id, std::string data) override;
-    void eventTextInputConfirm(int evt_id, std::string data) override;
-    void eventButtonPressed(int evt_id) override;
-    void eventSliderPosChanged(int evt_id, float f) override;
-    void eventFileOpen(int evt_id, PlatformNativePathString name, int importerIndex = -1) override;
-
     void updateEraserAndAlphaBlendButtons();
     void toggleEraser();
     void toggleAlphaBlendMode();
-    void updateMainEditorColor();
-    void updateMainEditorColorFromRGBSliders();
-    void updateColorModelSliders(std::string dontUpdate = "");
-    void updateMainEditorColorFromRGBTextBoxes();
-    void updateMainEditorColorFromHSVTextBoxes();
-    void setMainEditorColorHSV(double h, double s, double v);
-    void setMainEditorColorRGB(unsigned int col);
-    virtual void setMainEditorColorRGB(SDL_Color col, bool updateHSVSliders = true, bool updateRGBSliders = true, bool updateHSVTextBoxes = true, std::string dontUpdateThisColorModel = "");
-
-    void updateRGBTextBoxOnInputEvent(std::string data, uint8_t* value);
-    void updateHSVTextBoxOnInputEvent(std::string data, double* value);
 
     void pushLastColor(uint32_t col);
-    virtual void updateLastColorButtons();
-    void reloadColorLists();
-
-    void editorColorHSliderChanged(double h) {
-        currentH = h;
-        updateMainEditorColor();
-    }
-    void editorColorSVPickerChanged(double s, double v) {
-        currentS = s;
-        currentV = v;
-        updateMainEditorColor();
-    }
-
-#if _WIN32
-    void openOldWindowsColorPicker();
-#endif
+    void updateLastColorButtons() override;
 };
